requirements.txt
astroid==2.12.12
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
Flask==2.2.2
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
psycopg2-binary==2.9.4
pylint==2.15.5
python-dotenv==0.21.0
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
